Move TabNet to tab_model_base, it inherits from nn.Module, but we can't import nn.Module in tab_transformer_model.py.
